src/requirements-mac.txt
py2app==0.28.2
pyserial==3.5
pyyaml==6.0
numpy==1.23.1
scipy==1.8.1
PyQt6==6.3.1
matplotlib==3.5.2
